% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/behaviour_experiment.R
\name{behaviour_experiment}
\alias{behaviour_experiment}
\title{Simulate results of a behaviour experiment}
\usage{
behaviour_experiment(treatments, params)
}
\arguments{
\item{treatments}{Explanatory variable values in experimental treatments}

\item{params}{Parameters used by \emph{suitability} to generate environment relationship. Don't worry about this as this function is only called internally.}
}
\description{
This function simulates the outcome of a behavioural (stress-gulping) assay from a given input set of treatments
}
\author{
Martin Sullivan
}
