% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_death.R
\name{generate_death}
\alias{generate_death}
\title{Generate deaths (kill the minnow)}
\usage{
generate_death(prob)
}
\arguments{
\item{prob}{Probability of occurence. Generated by *suitability*}
}
\description{
This function generates a vector of deaths (binary variable) based on an input vector of probabilities of occurence
}
\author{
Martin Sullivan
}
