% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_gulp.R
\name{generate_gulp}
\alias{generate_gulp}
\title{Generate gulps}
\usage{
generate_gulp(prob, stock = 0)
}
\arguments{
\item{prob}{Probability of occurence. Generated by \emph{suitability}}

\item{stock}{Level of stocasticity (sd of rnorm)}
}
\description{
This function generates a vector of counts of stress gulps based on an input vector of probabilities of occurence
}
\author{
Martin Sullivan
}
