% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_presence.R
\name{generate_presence}
\alias{generate_presence}
\title{Generate presences}
\usage{
generate_presence(prob)
}
\arguments{
\item{prob}{Probability of occurence. Generated by *suitability*}
}
\description{
This function generates a vector of presences and absences based on an input sequence of probabilities of occurence
}
\author{
Martin Sullivan
}
