% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_minnow.R
\name{save_minnow}
\alias{save_minnow}
\title{Save the Marty Minnow}
\usage{
save_minnow()
}
\description{
Put your work into practice and see if you can save the Marty Minnow. This function takes interative input, and asks you to set the temperature and pH you recommend keeping your minnows at. It will then simulate how many minnows survive - have you saved the minnow, or is there something fishy with your advice?
}
\author{
Martin Sullivan
}
