% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suitability.R
\name{suitability}
\alias{suitability}
\title{Suitability}
\usage{
suitability(
  v1,
  v2,
  save.parms = FALSE,
  a1 = -10,
  a2 = 3,
  a3 = -0.3,
  a4 = 5,
  a5 = -0.6,
  a6 = -0.02,
  v2clamp = 2
)
}
\arguments{
\item{v1}{Environmental variable 1}

\item{v2}{Environmental variable 2}

\item{save.parms}{Do you want to save parameters? If so will output a list.}

\item{a1}{For me to know and you to find out, minnow marvelors}

\item{a2}{ditto}

\item{a3}{ditto}

\item{a4}{likewise}

\item{a5}{yep, see a1}

\item{a6}{ditto}

\item{v2clamp}{Documenting these would be giving too much away, sorry!}
}
\description{
Predict suitability given two environmental parameters. Can be set or will randomly generate.
}
\author{
Martin Sullivan
}
